gateone.js:  Fixed a bug where GateOne.User stuff wasn't working due to the 'prefix' variable being set incorrectly. gateone.py:  Fixed a bug where themes were not being synchronized with the client properly (it was using the modification time of the wrong file). gateone.py:  You can run Gate One out of the git clone directory again.  Installing via setup.py is not necessary.  This should close issue #214 (https://github.com/liftoff/GateOne/issues/214). gateone.py:  Changed the way 'gateone' settings are detected so as to avoid accidentally overwriting a 10server.conf that may simply have a syntax error (related to issue #213). app_terminal.py:  Fixed a bug where if you had a syntax error in your 50terminal.conf it would overwrite the file.  This should close issue #213 (https://github.com/liftoff/GateOne/issues/213). Themes:  Just a few tweaks here and there to make stuff look better.

+++ gateone/applications/terminal/app_terminal.py
@@ -1947,37 +1947,39 @@ def init(settings):
                 settings['*']['terminal'].update({key: value})
     if 'terminal' not in settings['*']:
         # Create some defaults and save the config as 50terminal.conf
-        from utils import settings_template
         settings_path = options.settings_dir
         terminal_conf_path = os.path.join(settings_path, '50terminal.conf')
-        # TODO: Think about moving 50terminal.conf template into the terminal
-        # application's directory.
-        template_path = os.path.join(
-            GATEONE_DIR, 'templates', 'settings', '50terminal.conf')
-        settings['*']['terminal'] = {}
-        # Update the settings with defaults
-        default_command = (
-            GATEONE_DIR +
-            "/applications/terminal/plugins/ssh/scripts/ssh_connect.py -S "
-            r"'%SESSION_DIR%/%SESSION%/%SHORT_SOCKET%' --sshfp "
-            r"-a '-oUserKnownHostsFile=\"%USERDIR%/%USER%/.ssh/known_hosts\"'")
-        settings['*']['terminal'].update({
-            'dtach': True,
-            'session_logging': True,
-            'session_logs_max_age': "30d",
-            'syslog_session_logging': False,
-            'commands': {
-                'SSH': default_command
-            },
-            'default_command': 'SSH'
-        })
-        new_term_settings = settings_template(
-            template_path, settings=settings['*']['terminal'])
-        with open(terminal_conf_path, 'w') as s:
-            s.write(_(
-                "// This is Gate One's Terminal application settings "
-                "file.\n"))
-            s.write(new_term_settings)
+        if not os.path.exists(terminal_conf_path):
+            from utils import settings_template
+            # TODO: Think about moving 50terminal.conf template into the
+            # terminal application's directory.
+            template_path = os.path.join(
+                GATEONE_DIR, 'templates', 'settings', '50terminal.conf')
+            settings['*']['terminal'] = {}
+            # Update the settings with defaults
+            default_command = (
+              GATEONE_DIR +
+              "/applications/terminal/plugins/ssh/scripts/ssh_connect.py -S "
+              r"'%SESSION_DIR%/%SESSION%/%SHORT_SOCKET%' --sshfp "
+              r"-a '-oUserKnownHostsFile=\"%USERDIR%/%USER%/.ssh/known_hosts\"'"
+            )
+            settings['*']['terminal'].update({
+                'dtach': True,
+                'session_logging': True,
+                'session_logs_max_age': "30d",
+                'syslog_session_logging': False,
+                'commands': {
+                    'SSH': default_command
+                },
+                'default_command': 'SSH'
+            })
+            new_term_settings = settings_template(
+                template_path, settings=settings['*']['terminal'])
+            with open(terminal_conf_path, 'w') as s:
+                s.write(_(
+                    "// This is Gate One's Terminal application settings "
+                    "file.\n"))
+                s.write(new_term_settings)
     term_settings = settings['*']['terminal']
     if options.kill:
         from utils import killall
